% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.bin.fac.R
\name{convert.bin.fac}
\alias{convert.bin.fac}
\title{Take a data.frame and convert any columns consisting of only 0's and 1's (binary) values, to a factor variable, to ease calculation and interpreations of subsequent analysis.}
\usage{
convert.bin.fac(x, ignore.missing = FALSE)
}
\arguments{
\item{x}{a data frame of patient characteristics from a study.}

\item{ignore.missing}{a scalar logical indicating on how to deal with missing data.  When true, the function will convert any column with 0's/1's/NA's into a factor, when false (default), NA's will prevent conversion.}
}
\value{
the same data.frame with updated columns
}
\description{
Take a data.frame and convert any columns consisting of only 0's and 1's (binary) values, to a factor variable, to ease calculation and interpreations of subsequent analysis.
}
\details{
This function convert any columns consisting of only 0's and 1's (binary) values, to a factor variable, to ease calculation and interpreations of subsequent analysis.
}
\examples{

set.seed(1)
N <- 100;
dataf <- data.frame(age=c(runif(N-1,0,100),NA),sex=sample(c(0,1,NA),N,replace=TRUE),cholesterol=rnorm(N,100,30))
summary(dataf)
out <-  convert.bin.fac(dataf)
summary(out)
out <- convert.bin.fac(dataf,ignore.missing=TRUE)
summary(out)
}
\author{
Jesse D. Raffa
}
\seealso{
\code{\link[knitr]{kable}} \code{\link[xtable]{xtable}} \code{\link[tableone]{CreateTableOne}}
}

