% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prop_by_level.R
\name{plot_prop_by_level}
\alias{plot_prop_by_level}
\title{Statified Proportion Plot with Confidence Intervals}
\usage{
plot_prop_by_level(dat, factor.var1, prop.var, alpha = 0.05,
  factor.var2 = NULL, ylab = "Proportion")
}
\arguments{
\item{dat}{a data frame of patient characteristics from a study.}

\item{factor.var1}{string indicating the first factor variable to stratify by (what will be on the x-axis)}

\item{prop.var}{string indicating the outcome variable to calculate the proportion of.  Must be binary or logical}

\item{alpha}{the confidence level to use for the confidence interval construction.  by default it uses 0.05 => 95\% confidence intervals.}

\item{factor.var2}{string indicating the second factor variable to stratify by (what will be designated by color), default is NULL which ignores this argument.}
}
\value{
a ggplot of the probability of the outcome stratified by factor.var1 and (optionally) factor.var2
}
\description{
Statified Proportion Plot with Confidence Intervals
}
\details{
This function produic
}
\examples{

set.seed(1)
require(ggplot2); require(dplyr)
N <- 100;
dataf <- data.frame(outcome=sample(c(0,1),N,replace=TRUE),age=c(runif(N-1,0,100),NA),sex=as.factor(sample(c(0,1,NA),N,replace=TRUE)),cholesterol=rnorm(N,100,30),rx=as.factor(sample(c(0,1),N,replace=TRUE)))
plot_prop_by_level(dataf,"rx","outcome",factor.var2="sex")
}
\author{
Jesse D. Raffa
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}

